requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.41
alembic==1.16.2
pydantic==2.11.7
email-validator==2.2.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.1.1
pydantic-settings==2.10.1
pillow==11.2.1
numpy==2.0.2
easyocr==1.7.2
PyMuPDF==1.26.1
pandas==2.3.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
gunicorn>=23.0.0